pyproject.toml
[project]
name = "cloudjack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.49",
    "google-cloud-secret-manager>=2.26.0",
    "google-cloud-storage>=3.9.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
